Keep-alive requirements
This removes the bad advice in PING about keeping connections alive and goes
for a more nuanced description.  The transport contains generic advice and a
recommendation that application protocols provide further guidance.  The HTTP
mapping specifies that the client is the one that performs keep-alives and that
it should only do so if it is waiting on responses from the server.

Closes #729.